First attempt to restructure repo for addon install in HASS.
