--- epson_projector_bridge/dockerfs/etc/services.d/epson_projector/run
@@ -1,3 +1,0 @@
-#!/command/with-contenv /bin/sh
-
-python /usr/src/app/asyncio_mqtt_based_client.py
