+++ epson_projector_mqtt_bridge/setup.cfg
