+++ repository.yaml
@@ -1,3 +1,3 @@
-name: Epson Projector Bridge
+name: Epson Projector to MQTT Bridge
 url: 'https://github.com/kjack9/epson_projector'
 maintainer: kjack9 <github@kellerg.com>
